+++ GorillaTemplate/Assets/_GorillaTemplate/Gorilla/UI/SeamlessRoomCodeDisplay.cs
@@ -1,32 +1,101 @@
+using Normal.Realtime;
 using Normal.SeamlessRooms;
 using UnityEngine;
+using UnityEngine.Serialization;
 
 namespace Normal.GorillaTemplate.UI {
     /// <summary>
     /// Displays the current room name for a <see cref="SeamlessRoomConnecter"/>.
     /// </summary>
     public class SeamlessRoomCodeDisplay : MonoBehaviour {
-        /// <summary>
-        /// The text field that will hold the name.
-        /// </summary>
-        [SerializeField]
-        private TMPro.TMP_Text _visual;
-
         [SerializeField]
         private SeamlessRoomConnecter _connecter;
 
-        private void Update() {
+        /// <summary>
+        /// A prefix to display before the name, ex "Current Room: ".
+        /// </summary>
+        [SerializeField]
+        private string _prefix;
 
+        /// <summary>
+        /// The text field that will hold the name.
+        /// </summary>
+        [FormerlySerializedAs("_visual")]
+        [SerializeField]
+        private TMPro.TMP_Text _nameLabel;
+
+        /// <summary>
+        /// The text field that will hold the disconnect event details (if any).
+        /// </summary>
+        [SerializeField]
+        private TMPro.TMP_Text _disconnectEventLabel;
+
+        /// <summary>
+        /// The text that describes the room.
+        /// </summary>
+        public string nameText {
+            get => _nameText;
+            private set {
+                _nameText = value;
+                if (_nameLabel != null) {
+                    _nameLabel.text = value;
+                }
+            }
+        }
+
+        private string _nameText;
+
+        /// <summary>
+        /// The text that describes the disconnect event details (if any).
+        /// </summary>
+        public string disconnectEventText {
+            get => _disconnectEventText;
+            private set {
+                _disconnectEventText = value;
+                if (_disconnectEventLabel != null) {
+                    _disconnectEventLabel.text = value;
+                }
+            }
+        }
+
+        private string _disconnectEventText;
+
+        private void Awake() {
+            _connecter.onWillConnect     += OnWillConnect;
+            _connecter.onDisconnectEvent += OnDisconnectEvent;
+            _connecter.realtime.didDisconnectFromRoomWithEvent += OnRealtimeDisconnectEvent;
+        }
+
+        private void OnDestroy() {
+            if (_connecter != null) {
+                _connecter.onWillConnect     -= OnWillConnect;
+                _connecter.onDisconnectEvent -= OnDisconnectEvent;
+
+                if (_connecter.realtime != null) {
+                    _connecter.realtime.didDisconnectFromRoomWithEvent -= OnRealtimeDisconnectEvent;
+                }
+            }
+        }
+
+        private void Update() {
             string description;
 
-            if (_connecter.roomInProgress == null) {
-                description = DisplayRealtimeStatus();
+            if (_connecter.TryGetComponent(out AutoReconnect autoReconnect) && autoReconnect.isReconnecting) {
+                if (autoReconnect.waitTime == 0f) {
+                    description = $"Reconnecting...";
+                } else {
+                    description = $"Reconnecting... ({Mathf.RoundToInt(autoReconnect.remainingTime)})";
+                }
             } else {
-                // A connection to a different room is in progress
-                description = DisplayRoomInProgressStatus();
+                if (_connecter.roomInProgress == null) {
+                    description = DisplayRealtimeStatus();
+                } else {
+                    // A connection to a different room is in progress
+                    description = DisplayRoomInProgressStatus();
+                }
             }
 
-            _visual.text = $"Current room: {description}";
+            nameText = $"{_prefix}{description}";
         }
 
         private string DisplayRealtimeStatus() {
@@ -38,8 +107,7 @@ namespace Normal.GorillaTemplate.UI {
             } else if (realtime.connecting) {
                 return "Connecting...";
             } else if (realtime.connected) {
-                // Use the room name
-                return realtime.room.name;
+                return RoomCodeDisplay.GetRoomDisplayName(realtime.room);
             } else {
                 return null;
             }
@@ -47,6 +115,23 @@ namespace Normal.GorillaTemplate.UI {
 
         private string DisplayRoomInProgressStatus() {
             return "Connecting...";
+        }
+
+        private void OnWillConnect() {
+            disconnectEventText = string.Empty;
+        }
+
+        private void OnDisconnectEvent(DisconnectEvent evt) {
+            if (evt is DisconnectCalledByLocalClient) {
+                return;
+            }
+
+            var details = evt == null ? "Unspecified disconnect reason" : evt.message;
+            disconnectEventText = $"{RoomCodeDisplay.GetRoomDisplayName(evt)}: {details}";
+        }
+
+        private void OnRealtimeDisconnectEvent(Realtime.Realtime realtime, DisconnectEvent evt) {
+            OnDisconnectEvent(evt);
         }
     }
 }