+++ GorillaTemplate/Assets/_GorillaTemplate/SeamlessRooms/SeamlessRoomConnecter.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using System;
+using System.Collections;
 using Normal.Realtime;
 using UnityEngine;
 
@@ -10,14 +11,36 @@ namespace Normal.SeamlessRooms {
     [RequireComponent(typeof(Realtime.Realtime))]
     public class SeamlessRoomConnecter : MonoBehaviour {
         /// <summary>
+        /// The quickmatch room group name to connect with.
+        /// </summary>
+        public string quickmatchGroupName = "default";
+
+        /// <summary>
+        /// When creating a new quickmatch room, this will be the max number of players of the new room.
+        /// </summary>
+        public int quickmatchRoomCapacity = 16;
+
+        /// <summary>
         /// The Realtime instance that this component is managing.
         /// </summary>
-        public Realtime.Realtime realtime => _realtime;
+        public Realtime.Realtime realtime {
+            get {
+                if (_realtime == null) {
+                    // In case this is called before Awake()
+                    return GetComponent<Realtime.Realtime>();
+                } else {
+                    return _realtime;
+                }
+            }
+        }
 
         /// <summary>
         /// The room that will be set on the Realtime once it's connected and ready.
         /// </summary>
         public Room roomInProgress => _roomInProgress;
+
+        public event Action onWillConnect;
+        public event Action<DisconnectEvent> onDisconnectEvent;
 
         /// <inheritdoc cref="realtime"/>
         private Realtime.Realtime _realtime;
@@ -48,6 +71,76 @@ namespace Normal.SeamlessRooms {
                 return;
             }
 
+            PrepareConnection(ref connectOptions);
+
+            // Tell the room to connect
+            _roomInProgress.Connect(roomName, connectOptions);
+        }
+
+        /// <inheritdoc cref="ConnectDirectlyToQuickmatchRoom(string)"/>
+        public void ConnectDirectlyToQuickmatchRoom(string roomGroupName, string roomCode, Room.ConnectOptions connectOptions = default) {
+            if (_realtime.disconnected == false && roomGroupName == _realtime.room.quickmatchRoomGroupName && roomCode == _realtime.room.quickmatchRoomCode) {
+                Debug.Log($"Already connecting or connected to {roomGroupName} - {roomCode}, ignoring the {nameof(ConnectDirectlyToQuickmatchRoom)} call");
+                return;
+            }
+
+            if (_roomInProgress != null && roomGroupName == _roomInProgress.quickmatchRoomGroupName && roomCode == _roomInProgress.quickmatchRoomCode) {
+                Debug.Log($"Already connecting to {roomGroupName} - {roomCode}, ignoring the {nameof(ConnectDirectlyToQuickmatchRoom)} call");
+                return;
+            }
+
+            PrepareConnection(ref connectOptions);
+
+            // Tell the room to connect
+            _roomInProgress.ConnectDirectlyToQuickmatchRoom(roomGroupName, roomCode, connectOptions);
+        }
+
+        /// <inheritdoc cref="Quickmatch"/>
+        public void ConnectToNextAvailableQuickmatchRoom(string roomGroupName, int capacity, Room.ConnectOptions connectOptions = default) {
+            if (_realtime.connecting && roomGroupName == _realtime.room.quickmatchRoomGroupName) {
+                Debug.Log($"Already connecting to quickmatch group {roomGroupName}, ignoring the {nameof(ConnectToNextAvailableQuickmatchRoom)} call");
+                return;
+            }
+
+            if (_roomInProgress != null && roomGroupName == _roomInProgress.quickmatchRoomGroupName) {
+                Debug.Log($"Already connecting to quickmatch group {roomGroupName}, ignoring the {nameof(ConnectToNextAvailableQuickmatchRoom)} call");
+                return;
+            }
+
+            PrepareConnection(ref connectOptions);
+
+            // Tell the room to connect
+            _roomInProgress.ConnectToNextAvailableQuickmatchRoom(roomGroupName, capacity, connectOptions);
+        }
+
+        /// <summary>
+        /// Connects to a room asynchronously.
+        /// Will seamlessly swap the room on the Realtime once the new room is connected and ready.
+        /// </summary>
+        public void ConnectToRoom(string roomName) {
+            // Connect using the default ConnectOptions
+            Connect(roomName);
+        }
+
+        /// <summary>
+        /// Connects to a quickmatch room asynchronously.
+        /// Will seamlessly swap the room on the Realtime once the new room is connected and ready.
+        /// </summary>
+        public void ConnectDirectlyToQuickmatchRoom(string roomCode) {
+            // Connect using the default ConnectOptions
+            ConnectDirectlyToQuickmatchRoom(quickmatchGroupName, roomCode);
+        }
+
+        /// <summary>
+        /// Connects to a room asynchronously using quickmatch.
+        /// Will seamlessly swap the room on the Realtime once the new room is connected and ready.
+        /// </summary>
+        public void Quickmatch() {
+            // Connect using the default options
+            ConnectToNextAvailableQuickmatchRoom(quickmatchGroupName, quickmatchRoomCapacity);
+        }
+
+        private void PrepareConnection(ref Room.ConnectOptions connectOptions) {
             // If we were still connecting to a different room, destroy it
             DestroyRoomIfNeeded();
 
@@ -61,17 +154,7 @@ namespace Normal.SeamlessRooms {
             // Subscribe to state changes
             _roomInProgress.connectionStateChanged += OnConnectionStateChanged;
 
-            // Tell the room to connect
-            _roomInProgress.Connect(roomName, connectOptions);
-        }
-
-        /// <summary>
-        /// Connects to a room asynchronously.
-        /// Will seamlessly swap the room on the Realtime once the new room is connected and ready.
-        /// </summary>
-        public void ConnectToRoom(string roomName) {
-            // Connect using the default ConnectOptions
-            Connect(roomName);
+            onWillConnect?.Invoke();
         }
 
         private void OnConnectionStateChanged(Room room, Room.ConnectionState previousState, Room.ConnectionState currentState) {
@@ -83,6 +166,8 @@ namespace Normal.SeamlessRooms {
 
                 _swapRoomsCoroutine = StartCoroutine(SwapRooms(room));
             } else if (currentState == Room.ConnectionState.Disconnected || currentState == Room.ConnectionState.Error) {
+                onDisconnectEvent?.Invoke(room.disconnectEvent);
+
                 // The connection did not succeed, so we destroy the new room and stay connected to the current room
                 DestroyRoomIfNeeded();
             }